requirements.txt
pyTelegramBotAPI>=4.14.0
python-dotenv
openai
pytz
schedule
gspread
google-auth-oauthlib
pandas
flask
psycopg2-binary
sqlalchemy
